Moves casl module into `/modules`
