+++ package.json
@@ -111,6 +111,7 @@
     "coverageDirectory": "../coverage",
     "testEnvironment": "node",
     "moduleNameMapper": {
+      "^@modules/(.*)$": "<rootDir>/src/modules/$1",
       "^@entities/(.*)$": "<rootDir>/src/entities/$1"
     }
   }
