+++ src/app.module.ts
@@ -13,7 +13,7 @@ import { GqlAuthGuard } from './auth/gql-auth.guard'
 import { AuthModule } from './auth/auth.module'
 import { CalendarEventModule } from './entities/calendarEvent/calendarEvent.module'
 import { CareerModule } from './entities/career/career.module'
-import { CaslModule } from './casl/casl.module'
+import { CaslModule } from './modules/casl/casl.module'
 import { DiscussionModule } from './entities/discussion/discussion.module'
 import { EventModule } from './entities/event/event.module'
 import { FollowModule } from './entities/follow/follow.module'
